requirements.txt
streamlit==1.51.0
google-generativeai==0.8.5
pymupdf==1.26.6
python-dotenv
pandas
